python_agent/requirements.txt
# Python 3.12+ - Finance NL Agent (LangGraph)
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0
httpx>=0.27.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
